pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robo_manip_baselines"
description = "Software that integrates various imitation learning methods and benchmark task environments to provide baselines for robot manipulation"
readme = "README.md"
authors = [
  {name = "Masaki Murooka", email = "m-murooka@aist.go.jp"},
  {name = "Tomohiro Motoda", email = "tomohiro.motoda@aist.go.jp"},
  {name = "Ryoichi Nakajo", email = "ryoichi-nakajo@aist.go.jp"},
]
maintainers = [
  {name = "Masaki Murooka", email = "m-murooka@aist.go.jp"},
]
requires-python = ">= 3.8"
dependencies = [
  "ipdb",
  "ipython",
  "psutil",
  "pyyaml",
  "tqdm",
  "dataclasses",
  "numpy",
  "h5py",
  "scipy>=1.7.0",
  "matplotlib>=3.3.4",
  "opencv-python",
  "ffmpeg-python",
  "imageio>=2.14.1",
  "videoio==0.3.0",
  "gymnasium==1.0.0",
  "mujoco==3.1.6",
  "pyspacemouse",
  "torch",
  "torchvision",
  "torchcodec",
  "tensorboard",
  "pynput>=1.8.1",
  "flax>=0.7.2",
]
dynamic = ["version"]

[project.optional-dependencies]
sarnn = [
  "scikit-learn>=1.3.0",
  "torchinfo",
  "gdown",
]
act = [
  "packaging",
  "pyquaternion",
]
mt-act = [
  "packaging",
  "pyquaternion",
  "lru_cache",
  "click",
  "sentence-transformers",
]
diffusion-policy = [
  "cffi",
  "zarr>=2.12.0",
  "numcodecs",
  "hydra-core",
  "einops",
  "dill",
  "diffusers==0.11.1",
  "robomimic==0.2.0",
  "huggingface_hub==0.21.4",
]
real-ur5e = [
  "ur_rtde",
  "pyrealsense2",
]
real-xarm7 = [
  "xArm-Python-SDK",
  "pyrealsense2",
]
benchmarks = [
  "ruamel.yaml",
  "python-pptx",
  "imageio-ffmpeg",
]

[tool.setuptools.packages.find]
include = ["robo_manip_baselines", "robo_manip_baselines.*"]
